Update for DialogFragment Time and Date Pickers.
Use text representations which are clickable to bring up the pickers,
respectively.
Add time and date to TextEditor to track the current time/date.
When setting the time/date, update the associated TextView to ensure data
is up to date.
scheduledTime and scheduledDate should be replaced with a single Date
object.  Text representations can simply pull from parts of the Calendar
object.